SCC reader refactoring (#63)
SCC reader refactoring
