+++ src/main/python/ttconv/scc/reader.py
@@ -37,8 +37,9 @@ from ttconv.scc.codes.control_codes import SccControlCode
 from ttconv.scc.codes.mid_row_codes import SccMidRowCode
 from ttconv.scc.codes.preambles_address_codes import SccPreambleAddressCode
 from ttconv.scc.codes.special_characters import SccSpecialAndExtendedCharacter
-from ttconv.scc.scc_elements import SccCaptionStyle, SccCaptionParagraph, SccCaptionContent, SccCaptionLineBreak, \
-  SccCaptionText
+from ttconv.scc.content import SccCaptionContent, SccCaptionLineBreak, SccCaptionText
+from ttconv.scc.paragraph import SccCaptionParagraph
+from ttconv.scc.style import SccCaptionStyle
 from ttconv.scc.time_codes import SccTimeCode, SMPTE_TIME_CODE_NDF_PATTERN, SMPTE_TIME_CODE_DF_PATTERN
 from ttconv.style_properties import StyleProperties
 
@@ -52,7 +53,6 @@ PARITY_BIT_MASK = 0b01111111
 class _SccContext:
   def __init__(self):
     self.div: Optional[Div] = None
-    # self.regions: List[Region] = []
     self.count: int = 0
     self.safe_area_x_offset: int = 0
     self.safe_area_y_offset: int = 0
@@ -94,14 +94,14 @@ class _SccContext:
     pac_indent = pac.get_indent()
 
     if self.current_caption.get_style() is SccCaptionStyle.PaintOn:
-      if self.current_caption.current_text and self.safe_area_y_offset + pac_row == self.current_caption.get_row_offset() + 1:
+      if self.current_caption.get_current_text() and self.safe_area_y_offset + pac_row == self.current_caption.get_row_offset() + 1:
         # Creates a new Paragraph if the new caption is contiguous (Paint-On)
         self.set_current_to_previous()
 
         self.current_caption = SccCaptionParagraph(self.safe_area_x_offset, self.safe_area_y_offset, SccCaptionStyle.PaintOn)
         self.init_current_caption(time_code)
 
-      elif len(self.previous_captions) > 0 and self.previous_captions[0].current_text \
+      elif len(self.previous_captions) > 0 and self.previous_captions[0].get_current_text() \
           and self.safe_area_y_offset + pac_row == self.previous_captions[0].get_row_offset():
         # Pushes and erases displayed row so that it can be replaced by current row (Paint-On)
         self.push_previous_caption(self.current_caption.get_begin())
@@ -111,7 +111,7 @@ class _SccContext:
     if self.current_caption.get_style() is SccCaptionStyle.RollUp:
       # Ignore PACs for rows 5-11, but get indent from PACs for rows 1-4 and 12-15. (Roll-Up)
       if pac_row in range(5, 12):
-        self.current_caption.set_current_text_offsets()
+        self.current_caption.apply_current_text_offsets()
         return
 
       self.current_caption.set_column_offset(pac_indent)
@@ -119,11 +119,11 @@ class _SccContext:
       self.current_caption.set_row_offset(pac_row)
       self.current_caption.set_column_offset(pac_indent)
 
-    self.current_caption.current_text.add_style_property(StyleProperties.Color, pac.get_color())
-    self.current_caption.current_text.add_style_property(StyleProperties.FontStyle, pac.get_font_style())
-    self.current_caption.current_text.add_style_property(StyleProperties.TextDecoration, pac.get_text_decoration())
+    self.current_caption.get_current_text().add_style_property(StyleProperties.Color, pac.get_color())
+    self.current_caption.get_current_text().add_style_property(StyleProperties.FontStyle, pac.get_font_style())
+    self.current_caption.get_current_text().add_style_property(StyleProperties.TextDecoration, pac.get_text_decoration())
 
-    self.current_caption.set_current_text_offsets()
+    self.current_caption.apply_current_text_offsets()
 
   def process_mid_row_code(self, mid_row_code: SccMidRowCode):
     """Processes SCC Mid-Row Code to map it to the model"""
@@ -131,23 +131,23 @@ class _SccContext:
       raise ValueError("No current SCC caption initialized")
 
     self.current_caption.new_caption_text()
-    self.current_caption.set_current_text_offsets()
+    self.current_caption.apply_current_text_offsets()
 
-    self.current_caption.current_text.add_style_property(StyleProperties.Color, mid_row_code.get_color())
-    self.current_caption.current_text.add_style_property(StyleProperties.FontStyle, mid_row_code.get_font_style())
-    self.current_caption.current_text.add_style_property(StyleProperties.TextDecoration, mid_row_code.get_text_decoration())
+    self.current_caption.get_current_text().add_style_property(StyleProperties.Color, mid_row_code.get_color())
+    self.current_caption.get_current_text().add_style_property(StyleProperties.FontStyle, mid_row_code.get_font_style())
+    self.current_caption.get_current_text().add_style_property(StyleProperties.TextDecoration, mid_row_code.get_text_decoration())
 
   def process_attribute_code(self, attribute_code: SccAttributeCode):
     """Processes SCC Attribute Code to map it to the model"""
-    if not self.current_caption or not self.current_caption.current_text:
+    if not self.current_caption or not self.current_caption.get_current_text():
       raise ValueError("No current SCC caption nor content initialized")
 
     if attribute_code.is_background():
-      self.current_caption.current_text.add_style_property(StyleProperties.BackgroundColor, attribute_code.get_color())
+      self.current_caption.get_current_text().add_style_property(StyleProperties.BackgroundColor, attribute_code.get_color())
     else:
-      self.current_caption.current_text.add_style_property(StyleProperties.Color, attribute_code.get_color())
+      self.current_caption.get_current_text().add_style_property(StyleProperties.Color, attribute_code.get_color())
 
-    self.current_caption.current_text.add_style_property(StyleProperties.TextDecoration, attribute_code.get_text_decoration())
+    self.current_caption.get_current_text().add_style_property(StyleProperties.TextDecoration, attribute_code.get_text_decoration())
 
   def process_control_code(self, control_code: SccControlCode, time_code: SccTimeCode):
     """Processes SCC Control Code to map it to the model"""
@@ -184,9 +184,9 @@ class _SccContext:
       self.push_previous_caption(time_code)
 
       self.current_caption = SccCaptionParagraph(self.safe_area_x_offset, self.safe_area_y_offset, SccCaptionStyle.RollUp)
-      self.current_caption.caption_contents = previous_last_lines
+      self.current_caption.set_contents(previous_last_lines)
       self.init_current_caption(time_code)
-      self.current_caption.set_roll_up_row_offsets()
+      self.current_caption.apply_roll_up_row_offsets()
 
     elif control_code is SccControlCode.EOC:
       # Display caption (Pop-On)
@@ -198,12 +198,12 @@ class _SccContext:
       if len(self.previous_captions) > 0:
         # Set line breaks depending on the position of the content
         last_text: Optional[SccCaptionText] = None
-        for (index, content) in enumerate(self.previous_captions[0].caption_contents):
+        for (index, content) in enumerate(self.previous_captions[0].get_contents()):
           if not isinstance(content, SccCaptionText):
             continue
 
-          if last_text and self.previous_captions[0].current_text.is_contiguous(last_text):
-            self.previous_captions[0].caption_contents.insert(index, SccCaptionLineBreak())
+          if last_text and self.previous_captions[0].get_current_text().is_contiguous(last_text):
+            self.previous_captions[0].insert_content(index, SccCaptionLineBreak())
 
           last_text = content
 
@@ -220,10 +220,16 @@ class _SccContext:
 
     elif control_code is SccControlCode.CR:
       # Roll the display up one row (Roll-Up)
+      # Not used in this implementation since this SCC reader does not really map the roll-up effect,
+      # but it erases the displayed paragraph and resets a new paragraph with the resulting rows.
       pass
 
     elif control_code is SccControlCode.DER:
       # Delete to End of Row (Paint-On)
+      # The DER may be issued from any point on a row to delete all displayable characters, transparent
+      # spaces, and mid-row codes from (and including) the current cell to the end of the row.
+      # Not used in this implementation since this SCC reader does not map the text overlapping into
+      # the model (i.e. a row is erased when a PAC is received, so before a new caption is written onto it).
       pass
 
   def process_text(self, word: str, time_code: SccTimeCode):
@@ -231,19 +237,19 @@ class _SccContext:
     if self.current_caption.get_style() is SccCaptionStyle.PaintOn:
       if word.startswith(" "):
         self.current_caption.new_caption_text()
-        self.current_caption.current_text.text += word
-        self.current_caption.set_current_text_offsets()
-        self.current_caption.current_text.set_begin(time_code)
+        self.current_caption.get_current_text().append(word)
+        self.current_caption.apply_current_text_offsets()
+        self.current_caption.get_current_text().set_begin(time_code)
         return
 
       if word.endswith(" "):
-        self.current_caption.current_text.text += word
+        self.current_caption.get_current_text().append(word)
         self.current_caption.new_caption_text()
-        self.current_caption.set_current_text_offsets()
-        self.current_caption.current_text.set_begin(time_code)
+        self.current_caption.apply_current_text_offsets()
+        self.current_caption.get_current_text().set_begin(time_code)
         return
 
-    self.current_caption.current_text.text += word
+    self.current_caption.get_current_text().append(word)
 
   def flush(self, time_code: SccTimeCode):
     """Flushes the remaining current caption"""
@@ -422,7 +428,7 @@ class SccLine:
 
 
 #
-# scc reader
+# SCC reader
 #
 
 def to_model(scc_content: str):
@@ -431,7 +437,7 @@ def to_model(scc_content: str):
   context = _SccContext()
   document = Document()
 
-  # safe area must be a 32x15 grid, that represents 80% of the root area
+  # Safe area must be a 32x15 grid, that represents 80% of the root area
   root_cell_resolution = CellResolutionType(rows=19, columns=40)
   document.set_cell_resolution(root_cell_resolution)
 
@@ -456,10 +462,5 @@ def to_model(scc_content: str):
     time_code = scc_line.to_model(context)
 
   context.flush(time_code)
-
-  # for region in context.regions:
-  #   if not region.get_doc():
-  #     region.set_doc(document)
-  #     document.put_region(region)
 
   return document
